add progress bar
