+++ README.md
@@ -4,49 +4,24 @@ A library to generate in batch diagram sequence using websequencediagrams
 
 ## Prerequisites
 
-* NodeJS[node](#https://nodejs.org)
+* NodeJS[node](https://nodejs.org)
 
 ## Installation
+
+```
+$ npm install
+```
 
 ## Get Started
 
+```
+$ dia generate <input_file> -e png
+```
+
 ## TODO List
 
-
-- [x] Working with custom errors
-    - [ ] Add Response for ALB and API Gateway
-- [ ] Benchmarking (Memory, CPU)
-    - [ ] Working with CheckPathVariables and GetPathVariable in Standard library
-    - [ ] More tests
-    - [ ] Working with pagination
-
-## Installation
-
-To install, use `go get` using go modules:
-
-```
-go get github.com/edermanoel94/rest-go@latest
-```
-
-## Example
-
-To install and build:
-
-```
-cd examples
-```
-
-And use go modules for install packages:
-
-```
-go mod tidy 
-```
-
-Then, to run:
-
-```
-go build
-```
+- [x] Progress bar
+- [ ] Filename output without space
 
 ## Contributing
